Re-organise public API (#99)
* Re-organise public API

* pre-commit fixes
