--- docs/api/v04/datasets.md
@@ -1,5 +1,0 @@
-# Dataset models
-
-These models represent OME-zarr datasets, which are zarr groups with specific attribute metadata.
-
-::: ome_zarr_models.v04
