+++ docs/api/v04/hcs.md
@@ -1,0 +1,7 @@
+# High Content Screening
+
+::: ome_zarr_models.v04.hcs
+
+## Plate metadata
+
+::: ome_zarr_models.v04.plate
