+++ docs/api/v04/image-label.md
@@ -1,0 +1,7 @@
+# Image Label
+
+::: ome_zarr_models.v04.image_label
+
+## Other types
+
+::: ome_zarr_models.v04.image_label_types
