--- docs/api/v04/metadata.md
@@ -1,47 +1,0 @@
-# Metadata models
-
-These are models of the metadata contained within OME-zarr datasets.
-
-## axes
-
-::: ome_zarr_models.v04.axes
-
-## coordinate_transformations
-
-::: ome_zarr_models.v04.coordinate_transformations
-
-## hcs
-
-::: ome_zarr_models.v04.hcs
-
-## image
-
-::: ome_zarr_models.v04.image
-
-## image_label
-
-::: ome_zarr_models.v04.image_label
-
-## labels
-
-::: ome_zarr_models.v04.labels
-
-## multiscales
-
-::: ome_zarr_models.v04.multiscales
-
-## omero
-
-::: ome_zarr_models.v04.omero
-
-## plate
-
-::: ome_zarr_models.v04.plate
-
-## well
-
-::: ome_zarr_models.v04.well
-
-## base
-
-::: ome_zarr_models.v04.base
