+++ docs/api/v04/other.md
@@ -1,0 +1,5 @@
+# Other
+
+## base
+
+::: ome_zarr_models.v04.base
