+++ docs/api/v04/well.md
@@ -1,0 +1,7 @@
+# Well
+
+::: ome_zarr_models.v04.well
+
+## Other types
+
+::: ome_zarr_models.v04.well_types
