+++ mkdocs.yml
@@ -62,8 +62,12 @@ nav:
   - Tutorial: tutorial.py
   - API reference:
       - v04:
-          - api/v04/datasets.md
-          - api/v04/metadata.md
+          - api/v04/hcs.md
+          - api/v04/image.md
+          - api/v04/image-label.md
+          - api/v04/labels.md
+          - api/v04/well.md
+          - api/v04/other.md
   - Contributing: contributing.md
 
 watch:
