--- src/ome_zarr_models/v04/__init__.py
@@ -1,7 +1,0 @@
-from ome_zarr_models.v04.hcs import HCS
-from ome_zarr_models.v04.image import Image
-from ome_zarr_models.v04.image_label import ImageLabel
-from ome_zarr_models.v04.labels import Labels
-from ome_zarr_models.v04.well import WellGroup
-
-__all__ = ["HCS", "Image", "ImageLabel", "Labels", "WellGroup"]
